revisar salida de contenedor

+++ src/main/java/com/appchoferes/nomina/models/lorasdb/dtos/ContenedorTipo1DTO.java
@@ -28,5 +28,5 @@ public class ContenedorTipo1DTO {
     private Integer ChoferID;
     private Integer ClienteID;
     private String Cliente;
-    private Long InventarioID;
+    private Integer InventarioID;
 }
